soc: qcom: ipa: client handle is valid in ipa_cfg_ep_status()
Like ipa_cfg_ep(), the only caller of ipa_cfg_ep_status() is
ipa_setup_sys_pipe(), and we know that the client handle will
be valid at the time of the call.  So there's no reason for
ipa_cfg_ep_status() to check that.

Removing that check makes ipa_cfg_ep_status() always return 0.
Change it so it has void return type, and update the caller so it
doesn't have to handle an error.

Signed-off-by: Alex Elder <elder@linaro.org>