soc: qcom: ipa: kill ipa_is_ep_support_flt()
All ipa_is_ep_support_flt() does is check to see if a bit is set in
the IPA context endpoint filter bitmask.  That is clear enough, and
ipa_is_ep_support_flt() doesn't really add much information to that.
So get rid of ipa_is_ep_support_flt() and open-code what it does in
the one place it's used.

Signed-off-by: Alex Elder <elder@linaro.org>