INTERNAL: drm/msm: Integrate interconnect API in DPU
The interconnect framework is designed to provide a
standard kernel interface to control the settings of
the interconnects on a SoC.

The interconnect API uses a consumer/provider-based model,
where the providers are the interconnect buses and the
consumers could be various drivers.

DPU is one of the interconnect consumers which uses the
interconnect APIs to get the path between endpoints and
set its bandwidth/latency/QoS requirements for the given
interconnected path.

Change-Id: Id4b0c0b526cf5fb287330acfdd997cfa4407da93
Signed-off-by: Sravanthi Kollukuduru <skolluku@codeaurora.org>