soc: qcom: ipa: num_desc is good in ipa_send()
There are three callers of ipa_send(), and all of them supply a
value that's guaranteed to be within the endpoint's TLV count.

  - ipa_send_cmd() always supplies 1 for num_desc.
  - ipa_send_cmd_timeout() always supplies 1 for num_desc.
  - ipa_tx_dp() checks whether the number of descriptors required to
    send the socket is more than the ipa_if_tlv value for the
    endpoint, and if so, returns an error.  So this too will always
    supply a num_desc value that's in range.

Replace the check for num_desc being in range with assertion, and
get rid of the error handling.

This makes local variable be used only once (in the assertion).
Get rid of that value, and call the function that previously
assigned its value directly in the assertion.

Signed-off-by: Alex Elder <elder@linaro.org>