soc: qcom: ipa: return proper errors in ipa_send()
If a memory allocation fails, the correct error value to return is
ENOMEM.  If a DMA mapping fails, it's less obvious, but ENOMEM is
pretty typical, and does communicate that the problem is memory
related.

Change ipa_send() so it returns ENOMEM if either of these types of
failures occur.

The only other failure that can occur is indicated by the return
value from gsi_queue_xfer().  If that produces an error, return that
error value rather than EFAULT.

The only callers of ipa_send() ignore the exact return value, so
this causes no change to caller behavior.

Signed-off-by: Alex Elder <elder@linaro.org>