soc: qcom: ipa: kill IPA_FIFO_ELEMENT_SIZE
IPA_FIFO_ELEMENT_SIZE is used in one place, to compute a size that's
used when allocating the memory for the event and channel rings for
the WWAN RX pipe.

Instead, use the actual GSI ring element size, and adjust the
multipler used so the computed result is the same.  To make it
more obvious the multipler is now different, rename it to be
IPA_APPS_WWAN_CONS_RING_COUNT.

Signed-off-by: Alex Elder <elder@linaro.org>