soc: qcom: ipa: drop some unnecessary casts
The type of the ifru_data element in the ifreq structure is a void
pointer (to user space).  This matches the type of the first argument
of copy_to_user() and the second argument of copy_from_user().  The
types match, so there's no need to type cast these values when used
in ipa_wwan_ioctl().  (The cast previously forced the user space
attribute to be ignored.)

Signed-off-by: Alex Elder <elder@linaro.org>