soc: qcom: ipa: DMA mode always supplies dst consumer
In ipa_cfg_ep_mode(), a WARN_ON() call is made to ensure the
destination in the ep_mode structure passed in is a consumer if
the mode is DMA.  Only one caller supplies IPA_DMA as its mode
(setting up APPS_CMD_PROD), and that provides a ep_mode.dst value of
APPS_LAN_CONS.  So the destination client supplied is always a
consumer.  Change the WARN_ON() to be an assertion, to communicate
what is expected.

Signed-off-by: Alex Elder <elder@linaro.org>