soc: qcom: ipa: rework gsi_ring_ctx->max_num_elem
The meaning of the max_num_elem field of the gsi_ring_ctx structure
is not completely obvious.  Its value is the index of the last entry
in the ring, which is *one less* than the maximum number of elements.

Change it to represent the number of elements in the ring buffer,
and rename it to be "count" to avoid any misinterpretation.

Signed-off-by: Alex Elder <elder@linaro.org>