soc: qcom: ipa: parse is non-null in ipahal_read_reg_n_fields()
There is a check in ipahal_read_reg_n_fields() to ensure the parse
pointer associated with the register is non-null.  It always is, and
always should be.  Replace the check with an assertion, to
communicate this assumption explicitly.

Here is an exhaustive verification that all current users of this
function pass a register that has a non-null construct pointer:
    - ipa_reset_with_open_aggr_frame_wa() passes IPA_ENDP_INIT_CTRL_n,
      which has parse function ipareg_parse_endp_init_ctrl_n().
    - ipa_set_flt_tuple_mask() passes IPA_ ENDP_FILTER_ROUTER_HSH_CFG_n,
      which has parse function ipareg_parse_hash_cfg_n().
    - ipa_set_rt_tuple_mask() passes IPA_ENDP_FILTER_ROUTER_HSH_CFG_n,
      which was just mentioned.
    - ipahal_read_reg_fields() passes the reg value it receives.
      The only one passed is IPA_ SHARED_MEM_SIZE, which has parse
      function ipareg_parse_shared_mem_size().

Signed-off-by: Alex Elder <elder@linaro.org>