soc: qcom: ipa: kill IPA_ENDP_INIT_HOL_BLOCK_* registers
The IPA_ENDP_INIT_HOL_BLOCK_EN_n and IPA_ENDP_INIT_HOL_BLOCK_TIMER_n
registers are never used in the IPA code, so get rid of their
definitions in ipahal_reg.h and ipahal_reg.c.  They have associated
construct functions that can be removed, and have a type that is no
longer needed.  A field of this type is never used, and can be
removed from the ipa_ep_context structure as well.

Signed-off-by: Alex Elder <elder@linaro.org>