soc: qcom: ipa: assert offload enable config is valid
In ipareg_construct_endp_init_cfg_n(), replace a WARN_ON() check
that the cs_offload_en value provided is valid with an assertion.

Ultimately this configuration comes from the four callers of
ipa_setup_sys_pipe().  Here are the values they provide:
  - setup_apps_cmd_prod_pipe() doesn't set it, so its value is 0,
    which corresponds to IPA_DISABLE_CS_OFFLOAD, which is valid.
  - setup_apps_lan_cons_pipe() sets IPA_ENABLE_CS_OFFLOAD_DL, which
    is valid.
  - handle_ingress_format() either doesn't set it (meaning it uses
    IPA_DISABLE_CS_OFFLOAD), or sets it to IPA_ENABLE_CS_OFFLOAD_DL.
    Both of these are valid.
  - handle_egress_format() either doesn't set it (meaning it uses
    IPA_DISABLE_CS_OFFLOAD), or sets it to IPA_ENABLE_CS_OFFLOAD_UL.
    Both of these are valid.

So all callers supply valid values.  The assertion communicates the
requirement for any future users of this function.  Since this is
now only used for an assertion, remove the debug message that gets
logged by cs_offload_en_valid() in case an invalid value is supplied.

Signed-off-by: Alex Elder <elder@linaro.org>