soc: qcom: ipa: ipa_desc payload and frag are independent
There are four types of IPA descriptor.  One, IPA_DATA_DESC is never
used as an IPA descriptor type (it's used only for NOP requests).
Each of the other three is assigned as a descriptor type in one
place in the code:
one location.
    - IPA_DATA_DESC_SKB is used in ipa_tx_dp() only for the linear
      (main) portion of a socket buffer being transmitted.  In this
      case, only fields "len" and "pyld" are filled in for this
      descriptor.  (If there are no fragments in the socket buffer,
      the "callback", "user1", and "user2" are also filled in.)
    - IPA_DATA_DESC_SKB_PAGED is used in ipa_tx_dp() only for any
      fragments in a socket buffer being transmitted.  Here, fields
      "frag" and "len" are filled in.  (For the last descriptor,
      the "callback", "user1", and "user2" are also filled in.)
    - IPA_IMM_CMD_DESC is used by ipa_desc_fill_imm_cmd() to fill a
      single descriptor.  In addition to the type, the "pyld", "len",
      and "opcode" are filled in.

This shows that either the payload pointer or the fragment pointer
in an IPA descriptor is filled in, but never both.  Enclose these
two fields in an anonymous union in the descriptor structure, so
this fact is obvious.

Signed-off-by: Alex Elder <elder@linaro.org>