soc: qcom: ipa: kill ipa_ep_cfg_hdr_ext->hdr_total_len_or_pad_offset
The hdr_total_len_or_pad_offset field of the ipa_ep_cfg_hdr_ext structure,
when set, is always assigned value 0.  If it's not set it is also 0, so
its value is always 0.  Remove this endpoint configuration option, and
have 0 always be used in the HDR_TOTAL_LEN_OR_PAD_OFFSET field of the
IPA_ENDP_INIT_HDR_EXT_n register.

Signed-off-by: Alex Elder <elder@linaro.org>