soc: qcom: ipa: kill ipa_ep_cfg_hdr_ext->ipa_ep_cfg_hdr_ext
The ipa_ep_cfg_hdr_ext field of the ipa_ep_cfg_hdr_ext structure,
when set, is always assigned value false.  If it's not set it is
also false, so its value is always false.  Remove this endpoint
configuration option, and have 1 (representing big endian) always be
used in the HDR_ENDIANNESS field of the IPA_ENDP_INIT_HDR_EXT_n
register.  Define a new enumerated type hdr_endianness that defines
symbolically the interpretation of this field's value.

Signed-off-by: Alex Elder <elder@linaro.org>