soc: qcom: ipa: kill ipa_ep_cfg_deaggr->packet_offset_valid
The packet_offset_valid field of the ipa_ep_cfg_deaggr structure,
when set, is only set to false.  This is no different from when it's
not set, it's value is always false.  Remove this endpoint
configuration option, and have 0 always be used when setting
the PACKET_OFFSET_VALID field of the ENDP_INIT_DEAGGR_n register.

Signed-off-by: Alex Elder <elder@linaro.org>