soc: qcom: ipa: interrupt is good in ipa_remove_interrupt_handler()
There are actually no callers of ipa_remove_interrupt_handler().
But if there were, we should assume that they provide a valid
interrupt value--the type of the argument is enough to describe
exactly the valid range of values expected.  Get rid of the range
check for the interrupt value passed.

Signed-off-by: Alex Elder <elder@linaro.org>