soc: qcom: ipa: irq_num is good in ipa_remove_interrupt_handler()
We can see by inspection that none of the entries ipa_irq_mapping[]
is too big.

And it would be an error for a caller to attempt to remove a handler
that had not already been added, and doing that will have verified
the interrupt number wasn't negative.

Remove the block of code that verified that the irq_num is in range.

Signed-off-by: Alex Elder <elder@linaro.org>