soc: qcom: ipa: use pointer in ipa_handle_interrupt()
There's no need to copy the contents of the interrupt info structure
for the interrupt in ipa_handle_interrupt().  Currently nothing ever
removes an IPA interrupt handler, so there's no risk it will change
(and even if there were, this code is wrong because it doesn't
adequately protect the access of the array against concurrent
modification).

Just use a pointer to the array entry.

Signed-off-by: Alex Elder <elder@linaro.org>