soc: qcom: ipa: kill INTERRUPT_WORKQUEUE_NAME
INTERRUPT_WORKQUEUE_NAME just defines the name used for the deferred
interrupt handling interrupt workqueue.  It's unlikely to ever
change, and it's easier to just see the name used than to have to
chase down the name somewhere else.  So just use the string itself
in the call to create_singlethread_workqueue() and get rid of the
symbol.

Signed-off-by: Alex Elder <elder@linaro.org>