soc: qcom: ipa: embed work struct in interrupt info
In principle, any IPA interrupt could occur interrupt context.  And
therefore every IPA interrupt type could require a work structure in
order to delay its handling.  Rather than allocating the work
structure (and other data) dynamically, embed a work structure in
each entry of the interrupt_info array.

This elminates the need for an atomic allocation in the interrupt
handler (which could fail), streamlines the code a bit, and removes
the need for the ipa_interrupt_work_wrap structure.

Signed-off-by: Alex Elder <elder@linaro.org>