soc: qcom: ipa: don't allocate suspend struct
When handling a suspend interrupt, we allocate a structure that's
saved as the interrupt data.  The structure contains only a 32-bit
unsigned value, so allocating a structure for that at runtime is
overkill.  Nothing else uses interrupt data.

Rather than allocating a structure, just pass the endpoint bit map
by saving its cast value to the interrupt_data pointer.  Then update
the suspend handler function to expect this, and don't have the
handler or deferred handler expect to free the pointer value.

Signed-off-by: Alex Elder <elder@linaro.org>