soc: qcom: ipa: kill ipa_active_clients_log_insert()
This is the first step toward removing the active clients log
functionlity.  First, stop adding anything to the log.  This is only
done in ipa_active_clients_log_mod().  As a result we can remove
ipa_active_clients_log_insert().  This removes the last place the
log_rdy field of the ipa_active_clients_log_ctx structure is ever
read, so it can be removed too.

Signed-off-by: Alex Elder <elder@linaro.org>