soc: qcom: ipa: kill ipa_write_64()
The reason ipa_write_64() was created was to allow a 64-bit value to
be written to an unaligned memory area.  The standard way of doing
this is using put_unaligned() (and ipa_write_64() is just a wrapper
for that function).  Get rid of ipa_write_64() and just use
put_unaligned() instead.

Signed-off-by: Alex Elder <elder@linaro.org>