soc: qcom: ipa: define event bitmap operations
Create a set of operations to encapsulate everything that happens
with the GSI context event bitmap:
  - gsi_evt_bmap_init() initializes the bitmap so all unsupported
    and reserved event ids are recorded as unavailable
  - gsi_evt_bmap_exit() reverses the effect of gsi_evt_bmap_init()
  - gsi_evt_bmap_alloc() finds the first 0 bit in evt_bmap, sets it
    (marks it as allocated), and returns the bit number as an event id.
  - gsi_evt_bmap_free() frees a previously allocated event id, by
    clearing its bit in the event bitmap.

Don't bother checking to see if we exhaust the number of available
events.  We allocate four channels currently, and our target
hardware supports more than that.  Add an assertion, so any future
attempt to allocate more than that will be noticed.

Reformat the gsi_alloc_evt_ring() a little bit (mostly white space
changes).

Signed-off-by: Alex Elder <elder@linaro.org>