soc: qcom: ipa: change ipa_setup_apps_pipes() return
Change the type of the return value of ipa_setup_apps_pipes() to be
an int rather than a long.  If it gets an error code from one of the
functions it calls, return that error code rather than manufacturing
some specific value.  The sole caller of ipa_setup_apps_pipes() only
checks whether the result is zero, so this has no visible effect.

Signed-off-by: Alex Elder <elder@linaro.org>