soc: qcom: ipa: assert proper alignment in gsi_alloc_evt_ring()
In gsi_alloc_evt_ring() there is a test to verify the result of a
DMA allocation satisfies hardware alignment constraints.  The DMA
allocator already satisfies that constraint, so change the test into
an assertion.

Specifically the requirement is that the DMA address is aligned on a
ann address aligned on a power-of-2 that is at least the requested
size.  The DMA allocator uses get_order(size) to determine the
number of pages required to satisfy the request, and passes that to
bitmap_find_free_region() to find an available region of DMA memory.
Both the size and the alignment of the result will be aligned on
a power-of 2 larger than the requested size.

Signed-off-by: Alex Elder <elder@linaro.org>