soc: qcom: ipa: ch_id is good in gsi_alloc_channel()
There is one path that leads to gsi_alloc_channel() being called:
    ipa_setup_sys_pipe()
      ipa_gsi_setup_channel()
        gsi_alloc_channel()

The ch_id value is based on the client id supplied by the caller of
ipa_setup_sys_pipe().  The four callers and their values supplied are:
  - setup_apps_cmd_prod_pipe(): IPA_CLIENT_APPS_CMD_PROD
  - setup_apps_lan_cons_pipe(): IPA_CLIENT_APPS_LAN_CONS
  - handle_ingress_format(): IPA_CLIENT_APPS_WAN_CONS
  - handle_egress_format(): IPA_CLIENT_APPS_WAN_PROD

From the client id, the ch_id value comes from the ipa_gsi_chan_num
field in the ipa_ep_configuration[] array.  For these four clients,
the ipa_gsi_ch_num are 4, 5, 6, and 3, respectively.  All of these
are less than the the number of channels supported by the SDM845 IPA
hardware.

Signed-off-by: Alex Elder <elder@linaro.org>