soc: qcom: ipa: assert proper alignment in gsi_alloc_channel()
This is just like the previous patch, but this time related to
channel ring DMA allocation.  A check in gsi_validate_channel_props()
verifies the DMA base address for the channel ring is properly
aligned.  But we know the DMA code will already satisfy that
requirement, so convert the test into an assertion.  Move the
assertion so it immediately follows the DMA allocation, in
gsi_alloc_channel().

Note that gsi_set_channel_cfg() also calls gsi_validate_channel_props().
This is called in two places:
  - ipa_reconfigure_channel_to_gpi() sets up a new properties
    structure, based on a channel's existing configuration.  But the
    dma memory region is simply copied from the original, so it will
    be exactly as valid as it was originally.
  - Similarly, ipa_restore_channel_properties() calls it, but the
    properties structure used will be one that was originally used
    to set up the channel, and that too is already known valid.

Signed-off-by: Alex Elder <elder@linaro.org>