soc: qcom: ipa: size is properly aligned in gsi_validate_channel_props()
In gsi_validate_channel_props() there is a check to see if the
memory size is at least 16 byte aligned.  That "16" should really
be "GSI_RING_ELEMENT_SIZE", and we already know this alignment
constraint is satisfied, because the caller computes it by
multiplying GSI_RING_ELEMENT_SIZE by the ring count.  So get rid of
that check.

The DMA memory region supplied for gsi_set_channel_cfg() will be the
same as the original region (as noted earlier).

Signed-off-by: Alex Elder <elder@linaro.org>