soc: qcom: ipa: phys_base won't cross 32-bit boundary
The final check in gsi_validate_channel_props() ensures that the
memory range used for the ring does not cross the 2^32 boundary.
That is, the last ring entry must be in the same 2^32 region as the
first.

This is also guaranteed by the DMA allocator, and already guaranteed
by this assertion:
        ipa_assert(!(props->mem.phys_base % roundup_pow_of_two(size)));
Here's why.

First, the memory size (with size > 0) will be a power of 2 number
of pages in size.  That's the way the DMA allocator works.  (Add an
assertion to that effect though.)

We also know that the allocator will ensure the physical base is
aligned on the same power-of-2 as the memory size.  So:
	alignment = roundup_pow_of_two(size) = mem_size
	phys_base % alignment = 0
This is also true:
	(phys_base + alignment) % alignment = 0
Which means:
	(phys_base + mem_size) % alignment = 0

We need to show that a multiple of 2^32 cannot lie within the memory
range.  I.e., this cannot be true:
	phys_base % 2^32 < 2^32 < (phys_base % 2^32) + mem_size

We know alignment (and mem_size) is a power-of-2, so the only way
2^32 could sit within that range is if 2^32 was less than the
alignment.  I.e., the only way the above could be true is if:
	2^32 < alignment	or	2^32 < mem_size

So as long as mem_size (and therefore size) is less than 2^32 the
allocated DMA range will satisfy the alignment constraint of not
crossing a 2^32 boundary.

The size value we're using is a 32-bit value.  So it cannot be
larger than 2^32.  Therefore the DMA memory range for the ring will
never straddle a 2^32 boundary.

Finally, he DMA memory region supplied for gsi_set_channel_cfg()
(the other user of gsi_validate_channel_props()) will be the same as
the original region (as noted earlier).

Signed-off-by: Alex Elder <elder@linaro.org>