soc: qcom: ipa: kill ipa_is_valid_ep()
The only caller of ipa_is_valid_ep() is ipa_handle_interrupt(),
which uses it when a SUSPEND_TX interrupt arrives as a optimization
to avoid calling the SUSPEND_TX handler function if none of the
endpoints in the mask are valid.  The handler, ipa_suspend_handler(),
also checks whether the endpoint is valid, so there's not much
benefit to the optimization.

Signed-off-by: Alex Elder <elder@linaro.org>