soc: qcom: ipa: use an ordered workqueue for ipa_interrupt_wq
The concurrency managed workqueue system has been around for a
*long* time.  Switch over to using the "new" interface to create the
IPA interrupt workqueue.

The old single-threaded workqueue ensured that only one instance of
the work was run at a time.  For backward compatibility a macro
implements create_singlethread_workqueue() using the new functions,
but it includes the WQ_MEM_RECLAIM flag.  We aren't doing anything
related to memory reclaim with this workqueue, so we shouldn't set
it.

To get the desired "in-order, only one at a time" behavior we use
alloc_ordered_workqueue(), and pass 0 for flags.

Signed-off-by: Alex Elder <elder@linaro.org>