soc: qcom: ipa: kill suspend_wa_lock
The suspend_wa_lock spinlock was used to protect potentially
concurrent execution of ipa_process_interrupts().  It might have
been possible because:
  - Normal interrupt handling queues ipa_interrupt_work_func(),
    which calls ipa_process_interrupts() in workqueue context.
  - The TX_SUSPEND interrupt workaround also queues work, which
    calls enable_tx_suspend_work_func() after a delay, and which
    also calls ipa_process_interrupts() in workqueue context.

But both work functions are queued on the ipa_interrupt_wq, which is
ordered.  As a result, the concurrency-managed workqueue system will
preclude these two functions executing at the same time.

So there's no need to use the suspend_wa_lock to ensure mutual
exclusion, and we can get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>