soc: qcom: ipa: hide struct ipa_sys_context
The only place fields within the ipa_sys_context structure are
referenced is within "ipa_dp.c".  Move its definition into that file
so it's no longer public, and just declare it in "ipa_i.h" as a
struct type.  This requires moving the definition of the ipa_repl_ctx
structure as well.

Signed-off-by: Alex Elder <elder@linaro.org>