soc: qcom: ipa: stash sys pointer in ipa_setup_sys_pipe()
An optimization for endpoint setup is to preserve certain field
values so they don't have to be reinitialized when an endpoint is
reused.  For the endpoint context, the only field that's preserved
is the sys pointer.  It's nice to know that only the sys field is
being preserved this way, so just copy it, zero the entire endpoint
structure, and then re-assign the sys pointer again afterward.  This
also removes the requirement that the sys field be last in the
structure.

Signed-off-by: Alex Elder <elder@linaro.org>