soc: qcom: ipa: desc and timeout are good in ipa_send_cmd_timeout()
The only caller of ipa_send_cmd_timeout() passes a non-null
descriptor pointer and a non-zero timeout value.  Remove the test
for those conditions, and assert the timeout is non-zero.

Get rid of a debug message, as well as a WARN() call.  All immediate
command descriptors are filled by ipa_desc_fill_imm_cmd(), which
will never fill in the callback or user1 fields.

Signed-off-by: Alex Elder <elder@linaro.org>