soc: qcom: ipa: kill ipa_wwan_change_mtu()
All ipa_wwan_change_mtu() does is is verifies the proposed MTU is
less than WWAN_DATA_LEN=2000 (returning EINVAL if so), and then sets
the device's mtu field.

This is unnecessary.  If the ndo_change_mtu function pointer is not
set for the network device, __dev_set_mtu() will set dev->mtu.
The caller of __dev_set_mtu() is dev_set_mtu(), and it checks the
proposed MTU against the device's max_mtu value, returning EINVAL if
it's too big.

So all we have to do is set netdev->max_mtu in ipa_wwan_setup(), and
the generic networking framework will handle MTU setting for us.

Signed-off-by: Alex Elder <elder@linaro.org>