soc: qcom: ipa: be extra safe about skb access
In ipa_wwan_xmit() the statistics for the netdevice are updated with
the length of the socket buffer *after* handing the skb off to
ipa_tx_dp().  This isn't really safe.  Capture the length before
that call and use that saved value to update the statistics.

Get rid of the "ret" local; it's not not needed.

Signed-off-by: Alex Elder <elder@linaro.org>