soc: qcom: ipa: assign packet length unconditionally
In ipa_rx_common() the RX packet length is assigned only if the size
supplied to the function is non-zero.

This is the only place that field is ever assigned, so upon entry
to this function the field will already have value 0 (having been
allocated by one of the replenish functions using kmem_cache_zalloc()).

Just assign the len field unconditionally.

Signed-off-by: Alex Elder <elder@linaro.org>