soc: qcom: ipa: stop using the "fast" replenish function
If there is more than one core available, a "fast" replenish
function is used to supply RX buffers for the APPS_WAN_CONS pipe.
The "fast" version attempts to use two cores concurrently to supply
these buffers--one allocating the buffers and a second to send them
to the IPA hardware.  There is a third, "recycle" method for
replenshing RX buffers.

The three replenish methods have their merits.  However:
  - The simple (non-"fast") one is adequate (though it might not
    perform as well as the "fast" one)
  - Simplicity is key for upstream review
  - This kind of optimization would be better added (upstream)
    later, along with some statistics that justify their use.

Finally, the practice elsewhere is to replenish RX buffers in the
NAPI poll function, which would be even simpler than any of these.

Signed-off-by: Alex Elder <elder@linaro.org>