soc: qcom: ipa: replenish handler will not be "fast"
In ipa_setup_sys_pipe() there is conditional code based on whether
the repelish handler function is ipa_fast_replenish_rx_cache().
It never will be, so we can get rid of this block of code.

This eliminates the last reference to ipa_fast_replenish_rx_cache(),
so remove that function.

Signed-off-by: Alex Elder <elder@linaro.org>