soc: qcom: ipa: kill ipa_sys_context->rcycl_list
The rcycl_list list of the ipa_sys_context structure gets
initialized, but nothing is ever added to it.  So it will be empty
when ipa_cleanup_rx() is called.  This means there's nothing to
clean up there, so a block of code can be removed.  This leaves no
more users of the rcycl_list field, so get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>