soc: qcom: ipa: kill ipa_sys_context->free_wrapper
The free_wrapper function pointer field of the ipa_sys_context
structure is only assigned the address of ipa_free_rx_wrapper().
Get rid of the field and call that function in the places the
pointer was used previously.  Move the definition of the function up
in its source file so it's defined before it's used.

Signed-off-by: Alex Elder <elder@linaro.org>