soc: qcom: ipa: assert non-empty list in ipa_wq_write_done_common()
It would be a serious error for the descriptor list to be empty when
ipa_wq_write_done_common() is called.  Assert it's not empty rather
than just silently ignoring the problem.

Nothing else will delete the entry from the list, so there's no need
to do this check while holding the spinlock.

Signed-off-by: Alex Elder <elder@linaro.org>