soc: qcom: ipa: kill two ipa_bug() calls in ipa_lan_rx_pyld_hdlr()
Get rid of two blocks of code that culminate in calling ipa_bug().

One verifies that the source and destination endpoint index values
supplied in a packet status structure are within range.  If they
are not, the hardware is misbehaving.  There's no reason why *this*
type of fault is any more or less likely than any other fault, so
there's no real point in trying to catch it if it occurs.

Similarly, the second triggers a BUG() if the packet length
described by a status packet exceeds the maximum.  That should never
happen.

Signed-off-by: Alex Elder <elder@linaro.org>