soc: qcom: ipa: zero payload sizes are not valid
The only TX operation that should use a zero-length payload is a
NOP.  And the GSI should never supply a (normally) completed RX
operation that has a zero size.

Replace two error checks for this condition with assertions in the
RX payload handling functions.  And add an assertion that the length
of the incoming skb is sufficient to hold a status packet in
ipa_lan_rx_cb() before calling ipahal_pkt_status_parse() with it.

Signed-off-by: Alex Elder <elder@linaro.org>