msm: ipa: use post-init interrupt from modem
Stop using a write to /dev/ipa for triggering the second-stage IPA
initialization.  Instead, start it upon receipt of a "post-init"
SMP2P interrupt issued by the modem when it has completed the IPA
firmware load and other early IPA hardware initialization.

Add a void reference to ipa_firmware_load(), so we can remove
that and other code in a separate step.

Signed-off-by: Alex Elder <elder@codeaurora.org>