msm: ipa: don't devm_kzalloc() for gsi_ctx
Currently, gsi_ctx is allocated using the device managed resource
interface, meaning if an error occurs, it will be freed when the
device is released.

However gsi_ctx is a global variable, and it will become a bad
pointer if such an error ever occurs.  In practice this isn't a
problem, but it's better to make pointers NULL when they are no
longer valid.

Allocate the gsi_ctx structure using kzalloc(), and free it if an
error occurs after it's allocated in gsi_init().  Set the global
pointer to NULL once it's free.

(There is currently no inverse for gsi_init().)

Signed-off-by: Alex Elder <elder@linaro.org>