msm: ipa: validate gsi-base properties
A DT resource can have 64-bit values for its start address and size.
But we require these values to be 32 bits only.  Currently the start
address is checked (asserted) to be in range but not the size.

Instead, check both values immediately after getting them, and
return an error rather than failing an assertion.

Signed-off-by: Alex Elder <elder@linaro.org>