msm: ipa: don't devm_request_irq() for GSI IRQ
Currently, the GSI IRQ is requested using the device managed resource
interface, meaning if an error occurs, it will be freed when the
device is released.

The IRQ number is associated with the GSI context, which is a
global, so freeing the device will leave the GSI context with
invalid content when the device is released.

Use request_irq() to register the GSI IRQ, and free_irq() in
gsi_deregister_device() to release it explicitly.

Signed-off-by: Alex Elder <elder@linaro.org>