msm: ipa: use dev_id value passed to ipa_isr()
When an IRQ is requested, an opaque "dev_id" value is provided that
will be passed to the interrupt handler function.  We provide the
GSI context pointer, but it's currently unused by gsi_isr() (though
there's an assertion it's what we think it is).

Have gsi_isr() use that value rather than the global gsi_ctx.

Signed-off-by: Alex Elder <elder@linaro.org>