msm: ipa: get rid of GSI register maximum parameter symbols
Each parameterized register offset macro has an additional symbol
defined that indicates the maximum allowed value for the macro's
parameter(s).  For example, for GSI_EE_N_CH_K_CNTXT_0_OFFS(k, n)
symbols GSI_EE_N_CH_K_CNTXT_0_MAX_K and GSI_EE_N_CH_K_CNTXT_0_MAX_N
define the maximum "k" and "n" values that can be provided to the
macro.

None of these values is enforced in code, so they are merely
informational.  Furthermore, almost all of these values are
the same (3 is almost always the MAX_N value, and the MAX_K value is
either 15 or 30).

It's just as useful (perhaps moreso) to just define these
constraints in a comment at the top of the file that defines the
macros.  Doing it this way makes explicit that they are all
generally subect to the same rules.

Signed-off-by: Alex Elder <elder@linaro.org>