msm: ipa: update ipa_wwan_ioctl() for flow control
The RMNET_IOCTL_FLOW_ENABLE and RMNET_IOCTL_FLOW_DISABLE both return
EINVAL if called.  Instead, return ENOTSUPP to indicate it's a
defined but not implemented request.

I'm not sure, but I *assume* the RMNET_IOCTL_FLOW_SET_HNDL request
has something to do with FLOW_ENABLE.  Rather than silently ignoring
this request if it's seen, indicate in the return value that it's
not supported.  This request (and the other two) have not been
observed in the ChromeOS environment.

Signed-off-by: Alex Elder <elder@linaro.org>