msm: ipa: kill local variable in ipa_wwan_ioctl_extended()
The "rc" local variable in ipa_wwan_ioctl_extended() holds a return
code.  It is returned at the end of the case statement--and nothing
else happens outside the case statement.  Have each case return
an explicit code immediately after handling the case.  This makes it
more obvious without looking ahead that we're done.

Signed-off-by: Alex Elder <elder@linaro.org>