msm: ipa: kill struct ipa_rmnet_mux_val
The only member of the ipa_rmnet_mux_val structure is the mux_id
(an unsigned integer) used for a given channel.  And the only use of
that structure is in an array in the rmnet_ipa_context structure.

Replace the array with an array of unsigned integers, and get rid of
the structure.  Rename the array "mux_id" to make it more obvious
its purpose.

Signed-off-by: Alex Elder <elder@linaro.org>