msm: ipa: only scan registered mux_ids for duplicates
Entries in the mux_id[] array are added, and never removed.  And the
only time the rmnet_count field (which defines how many entries are
valid) is updated is within ipa_wwan_add_mux_channel(), protected
by a mutex.

When checking to see if a mux_id has already been added, there's no
need to look at entries in the array that have not been recorded
yet.

Signed-off-by: Alex Elder <elder@linaro.org>