msm: ipa: kill rmnet_ipa_ctx->num_q6_rules
The num_q6_rules field of the rmnet_ipa_ctx structure is never set,
so it will have value 0.  The only place it's used is to set the
a7_ul_flt_set field of that structure if set.  Neither field is
otherwise used, so they can both just get removed.

Signed-off-by: Alex Elder <elder@linaro.org>