soc: qcom: ipa: change IPA compatible string
Rob Herring said that best practice is not to define generic node
types (like "qcom,ipa"), along with a more specific types or
versions for those nodes (like qcom,ipa-hw-ver) in Device Tree
files.  The compatible string should define both the type of the
node and its version.  In general, he said it was better to define
the version based on the SoC that contains the hardware rather than
based on the version of the component itself (i.e., the version of
IPA in the SDM845, not IPA version 3.5.1).

Change the IPA driver to match a Device Tree node with compatible
string "qcom,ipa-sdm845" rather than "qcom,ipa".

Signed-off-by: Alex Elder <elder@linaro.org>