soc: qcom: ipa: kill ipa_version_get()
The version of IPA hardware is now implied by a Device Tree node
having a matching compatible string.  There is no longer any need to
look up the IPA hardware version.

Get rid of ipa_version_get(), and the code that calls it.  That
eliminates the only use of ipa_hw_version enumerated type, so remove
its definition as well.

Signed-off-by: Alex Elder <elder@linaro.org>