arm64: dts: change IPA compatible string
Stop using the "qcom,ipa-hw-ver" property to define the version of
the hardware supported by the IPA driver.  Instead, we will encode
the hardware version in the IPA node compatible string.

The compatible string will now be "qcom,ipa-sdm845".

This obviates the need for the "qcom,ipa-hw-ver" property (and the
driver won't look for it), so remove this property.

Finally, the only reason to include header "qcom,ipa.h" was to get
the IPA hardware version symbols.  This is no longer needed, so stop
including it.

Signed-off-by: Alex Elder <elder@linaro.org>