msm: ipa: fix two BUILD_BUG_ON() calls
I find that the Chrome build environment complains about the use of
__builtin_constant_p(); maybe it's not present in their version of
GCC.  Anyway, that is not technically a requirement, so comment out
that line in __field_mask_check().

Also, someone at Google commented out another BUILD_BUG_ON() call in
gsi.c:command().  I wasn't sure why, but I now think it's because
the argument provided is not constant, or not known at compile time.
So just delete that test rather than commenting it out.  It's a
little paranoid anyway.

Signed-off-by: Alex Elder <elder@linaro.org>