soc: qcom: ipa: kill struct ipahal_fltrt_obj
The rule_id_bit_len field of the ipahal_fltrt_obj structure is
assigned a constant value, and it's only used once (inside the file
that defines the field value).  Get rid of the field, and just use
the constant in the place it was previously used.

This makes the ipahal_fltrt_obj structure unused, so it can be
removed.  Move some of the comments explaining what it was used for
up above the definitions of the constants used in its place.

Signed-off-by: Alex Elder <elder@linaro.org>