soc: qcom: ipa: kill ipahal_get_hw_tbl_hdr_width()
Rather than hiding it behind a function, just expose the width of an
IPA filter table in "ipahal.h".  It's a constant, and until we
support different hardware (with a different table entry size)
the value of hiding this isn't very large.  It will allow us to
simplify the code a but further.

Signed-off-by: Alex Elder <elder@linaro.org>