soc: qcom: ipa: remove some unnecessary checks
The first check in config_valid() guarantees IPA_HW_TBL_HDR_WIDTH is
non-zero.  Both IPA_MEM_V4_RT_NUM_INDEX and IPA_MEM_V6_RT_NUM_INDEX
are similarly guaranteed to be non-zero.  Therefore their products
will be non-zero.

If IPA_MEM_V4_RT_HASH_SIZE were 0, this check would fail:
        BUILD_BUG_ON(IPA_MEM_V4_RT_NUM_INDEX * IPA_HW_TBL_HDR_WIDTH
                     > IPA_MEM_V4_RT_HASH_SIZE);
The same is true for similar checks with the non-hashed route table
size, as well as for similar checks for IPv6.

As a result, this check is redundant:
	BUILD_BUG_ON(!IPA_MEM_V4_RT_HASH_SIZE);
The other three similar checks are redundant as well.  All four of
these can therefore be removed.

By similar reasoning, the check to ensure IPA_MEM_V4_FLT_HASH_SIZE
(and other filter table sizes) is non-zero can be removed because
they are redundant.  Even if the number of filterable pipes is zero,
the number of entries in the filter table will be non-zero (because
of the extra entry holding the bitmask).  So the product of the
number of entries times the (non-zero) size of an entry will be
non-zero.

Signed-off-by: Alex Elder <elder@linaro.org>